Moved the logic to fetch food options out of gatsby-node.js and into it's own custom plugin called source-food-options
